+++ src/Shortcode/Gamecenter.php
@@ -4,7 +4,6 @@ namespace NovaStats\Gamecenter\Shortcode;
 
 defined('ABSPATH') || exit;
 
-use NovaStats\Gamecenter\Constant\PluginConstants;
 use NovaStats\Gamecenter\Helper\TemplateEngine;
 use NovaStats\Gamecenter\Model\Settings;
 use NovaStats\Gamecenter\Service\VueService;
@@ -32,7 +31,6 @@ class Gamecenter
         add_shortcode($name, function (array $atts): string {
             $data = [
               'sessionData' => $this->settings->getSessionInitialData(),
-              PluginConstants::INITIAL_DATA_KEY_GAMECENTER_BASE_URL => $this->settings->getGamecenterBaseUrl(),
             ];
 
             ob_start();